[Clean] Split some codes into several files
- Split some codes into several files in order to improve readability
