+++ src/file-utilities.ts
@@ -1,5 +1,5 @@
 // =============================================================================================================================
-// SRC - URI GENERATOR
+// SRC - FILE UTILITIES
 // =============================================================================================================================
 import * as mime from "mime";
